requirements.txt
anyio==3.7.0
asyncclick==8.1.3.4
blinker==1.6.2
click==8.1.3
Flask==2.3.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.3
pydantic==1.10.9
python-kasa==0.5.1
sniffio==1.3.0
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
